fix: creating query
