+++ src/page-modules/assistant/index.ts
@@ -1,9 +1,6 @@
 export * from './types';
 export { parseTripQuery } from './utils';
-export {
-  fetchFromToTripQuery,
-  type FromToTripQuery,
-} from './trip-query-fetcher';
+export { fetchFromToTripQuery } from './trip-query-fetcher';
 export type { AssistantLayoutProps } from './layout';
 export { default as AssistantLayout } from './layout';
